updates to routing
