+++ pages/studentPortal/quizcomponent/LangQuizRoute.tsx
@@ -1,4 +1,4 @@
-import { useState } from "react";
+import { useEffect, useState } from "react";
 import LangResults from "../../../components/quizzes/langQuiz/LangResults";
 import BluePrint from "../../../components/quizzes/shared/BluePrint";
 import SkillSelections from "../../../components/quizzes/shared/SkillSelections";
@@ -11,8 +11,15 @@ export enum Stage {
   BLUEPRINT,
   RESULTS,
 }
+
+type QuizSelection = {
+  name: string;
+  result?: string;
+  weight?: number;
+};
+
 const LangQuiz = () => {
-  const selectFields = [
+  const quiz: QuizSelection[][] = [
     [
       { name: "Learn an in-demand skill" },
       { name: "Work in tech" },
@@ -23,28 +30,52 @@ const LangQuiz = () => {
       { name: "Start a side hustle" },
     ],
     [
-      { name: "Front end development", lang: "JavaScript", weight: 1 },
-      { name: "Back end development", lang: "JavaScript", weight: 1 },
-      { name: "Game development", lang: "JavaScript", weight: 1 },
-      { name: "Mobile development", lang: "JavaScript", weight: 1 },
-      { name: "Data analytics", lang: "Python", weight: 1 },
-      { name: "Product management", lang: "JavaScript", weight: 1 },
-      { name: "Digital marketing", lang: "JavaScript", weight: 1 },
-      { name: "UX/UI design", lang: "JavaScript", weight: 1 },
-      { name: "I don't know / Not sure yet", lang: "JavaScript", weight: 1 },
+      { name: "Front end development", result: "JavaScript", weight: 1 },
+      { name: "Back end development", result: "JavaScript", weight: 1 },
+      { name: "Game development", result: "JavaScript", weight: 1 },
+      { name: "Mobile development", result: "JavaScript", weight: 1 },
+      { name: "Data analytics", result: "Python", weight: 1 },
+      { name: "Product management", result: "JavaScript", weight: 1 },
+      { name: "Digital marketing", result: "JavaScript", weight: 1 },
+      { name: "UX/UI design", result: "JavaScript", weight: 1 },
+      { name: "I don't know / Not sure yet", result: "JavaScript", weight: 1 },
     ],
     [
-      { name: "Websites", lang: "JavaScript", weight: 1 },
-      { name: "Storefront", lang: "HTML/CSS", weight: 1 },
-      { name: "Mobile apps", lang: "JavaScript", weight: 1 },
-      { name: "Games", lang: "JavaScript", weight: 1 },
-      { name: "Email", lang: "HTML/CSS", weight: 1 },
-      { name: "Tools to automate your life", lang: "JavaScript", weight: 1 },
-      { name: "I don't know / Not sure yet", lang: "JavaScript", weight: 1 },
+      { name: "Websites", result: "JavaScript", weight: 1 },
+      { name: "Storefront", result: "HTML/CSS", weight: 1 },
+      { name: "Mobile apps", result: "JavaScript", weight: 1 },
+      { name: "Games", result: "JavaScript", weight: 1 },
+      { name: "Email", result: "HTML/CSS", weight: 1 },
+      { name: "Tools to automate your life", result: "JavaScript", weight: 1 },
+      { name: "I don't know / Not sure yet", result: "JavaScript", weight: 1 },
     ],
   ];
 
-  const selectFieldName = selectFields.map((outerItem) => {
+  type QuizContext = {
+    title: string;
+    body: string;
+  };
+
+  const context: QuizContext[] = [
+    {
+      title: "What Coding Language Should I Learn First?",
+      body: "Take this free quiz to find out what coding language you should learn first.",
+    },
+    {
+      title: "Why do you want to learn coding?",
+      body: "Select all that apply.",
+    },
+    {
+      title: "What field of work are you interested in?",
+      body: "Select 1-3 choices.",
+    },
+    {
+      title: "What are you interested in building?",
+      body: "Select all that apply.",
+    },
+  ];
+
+  const selectFieldName = quiz.map((outerItem) => {
     return outerItem.map((innerItem) => {
       return innerItem.name;
     });
@@ -52,10 +83,50 @@ const LangQuiz = () => {
 
   const [stageResponses, setStageResponses] = useState([]);
 
-  const [score, setScore] = useState(0);
+  const [score, setScore] = useState({
+    JavaScript: 0,
+    HTMLCSS: 0,
+    Python: 0,
+  });
 
-  // create results state object that
-  // create custom type -- based on schema type in database
+  const finalScore = (userSelections: string | any[]) => {
+    for (let i = 0; i < userSelections.length; i++) {
+      if (userSelections[i] && userSelections[i].result === "JavaScript") {
+        setScore((score) => ({
+          ...score,
+          JavaScript: score.JavaScript + userSelections[i].weight,
+        }));
+      }
+      if (userSelections[i] && userSelections[i].result === "HTMLCSS") {
+        setScore((score) => ({
+          ...score,
+          HTMLCSS: score.HTMLCSS + userSelections[i].weight,
+        }));
+      }
+      if (userSelections[i] && userSelections[i].result === "Python") {
+        setScore((score) => ({
+          ...score,
+          Python: score.Python + userSelections[i].weight,
+        }));
+      }
+    }
+  };
+  const userSelections: QuizSelection[] = [];
+  const productResult = () => {
+    for (let i = 0; i < stageResponses.length; i++) {
+      const check = quiz[i].filter((item) =>
+        stageResponses[i].includes(item.name)
+      );
+      userSelections.push(...check);
+    }
+    finalScore(userSelections);
+    console.log(userSelections);
+  };
+
+  useEffect(() => {
+    productResult();
+  }, []);
+
   const [stage, setStage] = useState(Stage.START);
   const handleNextClick = () => {
     setStage((prevStage) => prevStage + 1);
@@ -63,17 +134,15 @@ const LangQuiz = () => {
   const handleBackClick = () => {
     setStage((prevStage) => prevStage - 1);
   };
-  // Render the appropriate component based on the stage
+
   const renderStage = () => {
     switch (stage) {
       case Stage.START:
         return (
           <StartQuiz
             onNextClick={handleNextClick}
-            title={"What Coding Language Should I Learn First?"}
-            body={
-              "Take this free quiz to find out what coding language you should learn first."
-            }
+            title={context[Stage.START].title}
+            body={context[Stage.START].body}
           />
         );
       case Stage.LEARNCODING:
@@ -81,10 +150,10 @@ const LangQuiz = () => {
           <SkillSelections
             onNextClick={handleNextClick}
             onBackClick={handleBackClick}
-            title={"Why do you want to learn coding?"}
-            body={"Select all that apply."}
+            title={context[Stage.LEARNCODING].title}
+            body={context[Stage.LEARNCODING].body}
             progress={20}
-            selections={selectFieldName[0]}
+            selections={selectFieldName[Stage.LEARNCODING - 1]}
             setStageResponses={setStageResponses}
             currentStage={Stage.LEARNCODING}
           />
@@ -94,10 +163,10 @@ const LangQuiz = () => {
           <SkillSelections
             onNextClick={handleNextClick}
             onBackClick={handleBackClick}
-            title={"What field of work are you interested in?"}
-            body={"Select 1-3 choices."}
+            title={context[Stage.WORKFIELDS].title}
+            body={context[Stage.WORKFIELDS].body}
             progress={50}
-            selections={selectFieldName[1]}
+            selections={selectFieldName[Stage.WORKFIELDS - 1]}
             setStageResponses={setStageResponses}
             currentStage={Stage.WORKFIELDS}
           />
@@ -107,10 +176,10 @@ const LangQuiz = () => {
           <SkillSelections
             onNextClick={handleNextClick}
             onBackClick={handleBackClick}
-            title={"What are you interested in building?"}
-            body={"Select all that apply."}
+            title={context[Stage.BUILDTARGET].title}
+            body={context[Stage.BUILDTARGET].body}
             progress={80}
-            selections={selectFieldName[2]}
+            selections={selectFieldName[Stage.BUILDTARGET - 1]}
             setStageResponses={setStageResponses}
             currentStage={Stage.BUILDTARGET}
           />
@@ -131,6 +200,7 @@ const LangQuiz = () => {
   return (
     <div>
       <p>{JSON.stringify(stageResponses)}</p>
+      <p>{context}</p>
       {renderStage()}
     </div>
   );
